start on recursive function WIP
Validate board wip

+++ vector2.go
@@ -1,6 +1,6 @@
 package main
 
 type Vector2 struct {
-	X int8
-	Y int8
+	X int
+	Y int
 }
